requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.1
billiard==3.6.3.0
celery==4.4.3
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
Django==3.0.6
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
isort==4.3.21
kombu==4.6.9
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipenv==2020.6.2
pylint==2.5.2
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.1.0
